+++ staging/src/k8s.io/client-go/tools/leaderelection/resourcelock/configmaplock.go
@@ -23,7 +23,7 @@ import (
 
 	"k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	corev1client "k8s.io/kubernetes/pkg/client/clientset_generated/clientset/typed/core/v1"
+	corev1client "k8s.io/client-go/kubernetes/typed/core/v1"
 )
 
 // TODO: This is almost a exact replica of Endpoints lock.
