+++ test/test_owners.csv
@@ -645,7 +645,7 @@ k8s.io/kubernetes/pkg/apis/rbac/validation,erictune,0,
 k8s.io/kubernetes/pkg/apis/storage/validation,caesarxuchao,1,
 k8s.io/kubernetes/pkg/auth/authorizer/abac,liggitt,0,
 k8s.io/kubernetes/pkg/client/chaosclient,deads2k,1,
-k8s.io/kubernetes/pkg/client/leaderelection,xiang90,1,
+k8s.io/client-go/tools/leaderelection,xiang90,1,
 k8s.io/kubernetes/pkg/client/legacylisters,jsafrane,1,
 k8s.io/kubernetes/pkg/client/listers/batch/internalversion,mqliang,0,
 k8s.io/kubernetes/pkg/client/listers/extensions/internalversion,eparis,1,
